seminars/manim/requirements.txt
manim>=0.18,<0.20
imagehash>=4.3
Pillow>=10.3
ffmpeg-python>=0.2
python-frontmatter>=1.1
jsonschema>=4
pytest>=8
